Struct xmpp_client::Client

pub struct Client { /* fields omitted */ }

Client struct deal with connection and message sharing between server and client

Methods

impl Client

fn connect(stream: TcpStream,
config: XMPPConfig,
credentials: Option<Credentials>)
-> Box<Future<Item=Client, Error=Error>>

Connect a client to a server

fn close(&self)

ok

fn send_presence(&self) -> Box<Future<Item=(), Error=Error>>

Send the first presence

fn get_jid(&self) -> Credentials

Return the current Credentials

fn send(&mut self, f: Event) -> Box<Future<Item=(), Error=Error>>

Send an event to the stream

TODO: need future here

fn shutdown(&mut self) -> Result<Async<()>, Error>

ok

fn send_ping(&mut self, ping: &mut Ping) -> Receiver<Event>

Send a ping request

fn handle(&mut self) -> Box<Future<Item=Consumer, Error=Error>>

Create a consumer and return a stream that Poll every Event in the stream